requirements.txt
PyMuPDF==1.24.2
markdown-it-py[plugins]==3.0.0
